% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{dilate}
\alias{dilate}
\title{Dilate terrain}
\usage{
dilate(r, w = 3)
}
\arguments{
\item{r}{\code{terra::SpatRaster} object containing a digital elevation model.}

\item{w}{integer specifying the window size used to dilate the terrain. Larger values are slower but produce more dramatic effects.}
}
\description{
Widen terrain features with a local maximum filter.
}
\details{
This simple function wraps a call to terra::focal with a maximum filter of the desired window size.
It can be used to quickly widen ridges and thin valleys. As a general rule, this function and its inverse \code{pinch()}
are best used early in a workflow prior to applying erosion algorithms.
}
