% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosion.R
\name{incise_flow}
\alias{incise_flow}
\title{Carve out drainage channels}
\usage{
incise_flow(r, log_base = 1.5, threshold = 100, blur = T)
}
\arguments{
\item{r}{\code{terra::SpatRaster} object containing a digital elevation model.}

\item{log_base}{numeric, the base of the logarithm used to transform the flow accumulation counts.}

\item{threshold}{integer, minimum number of upstream cells required to gouge a channel.}

\item{blur}{TRUE (default) or FALSE. Indicates whether the flow accumulation map should be blurred with a mean filter.}
}
\description{
Lower terrain according to the logarithm of flow accumulation.
}
\details{
This algorithm is very fast and works to create passable rills and canyons. However, it is prone to generating straight lines in flat areas and abruptly terminating canyons in basins with no outlet.
Both issues can be ameliorated by running one or more iterations of \code{worldbuildeR::add_noise()} followed by a basin-filling algorithm such as \code{flowdem::fill()}.
}
