% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{pinch}
\alias{pinch}
\title{Pinch terrain}
\usage{
pinch(r, w = 3)
}
\arguments{
\item{r}{\code{terra::SpatRaster} object containing a digital elevation model.}

\item{w}{integer specifying the window size used to pinch the terrain. Larger values are slower but produce more dramatic effects.}
}
\description{
Narrow terrain features with a local minimum filter.
}
\details{
This simple function wraps a call to terra::focal with a minimum filter of the desired window size.
It can be used to quickly thin ridges and widen valleys. As a general rule, this function and its inverse \code{dilate()}
are best used early in a workflow prior to applying erosion algorithms.
}
