% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosion.R
\name{rainfall_erode}
\alias{rainfall_erode}
\title{Erode with simulated rainfall agents}
\usage{
rainfall_erode(
  r,
  precipitons = as.integer(terra::ncell(r)/10),
  frac = 0.33,
  rainfall = terra::rast(r, vals = 1)
)
}
\arguments{
\item{r}{\code{terra::SpatRaster} object containing a digital elevation model.}

\item{precipitons}{Integer. Number of agents to simulate.}

\item{frac}{Numeric, in (0,0.5].}

\item{rainfall}{\code{terra::SpatRaster} object containing relative precipitation weights, used to weight the random start locations of the precipiton agents.}
}
\description{
This function iteratively
}
\details{
The precipiton erosion algorithm randomly drops a simulated agent on the landscape (optionally weighted, e.g. by annual rainfall).
The agent moves down the path of steepest slope, transferring \code{frac*(upslope_elevation - downslope_elevation)} from the upslope cell to the
downslope cell.

This method of simulating erosion is unfortunately much slower than \code{worldbuildeR::incise_flow()}. However,
it has the advantage of producing somewhat more organic-looking results and
not requiring filled drainage basins to function correctly.

Note that even if a \code{rainfall} raster is provided, the odds of dropping an agent over negative elevations will be set to zero.
}
