% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{render_terrain}
\alias{render_terrain}
\title{Render a DEM}
\usage{
render_terrain(r, shaded = T, flat_water = F)
}
\arguments{
\item{r}{\code{terra::SpatRaster} object containing a digital elevation model}

\item{shaded}{TRUE (default) or FALSE. Indicates whether to plot the DEM with hillshading as calculated by \code{terra::shade()}}

\item{flat_water}{TRUE or FALSE (default). Indicates whether to display a bathymetric color gradient for elevations below 0. If TRUE, all elevations below 0 will be displayed as a uniform shade of blue.}
}
\description{
Display a shaded relief map of a DEM with elevation and water level appropriately color-coded.
}
\details{
This function serves as a convenient wrapper for \code{terra::plot()}. It automatically generates and overlays a hillshade layer calculated using \code{terra::shade()}
and \code{terra::terrain()}, and it shades terrain elevation using Wikipedia's topographic palette 2.0.
}
