% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldbuildeR-package.R
\docType{package}
\name{worldbuildeR-package}
\alias{worldbuildeR}
\alias{worldbuildeR-package}
\title{worldbuildeR: Tools for Creating Speculative Geographies.}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
